fix #5, base64 secret was parsed with an ending \n
When I was executing the test suite I encountered the same issue as described in issue #5. Base64 secret was parsed with an ending \n.
I don't know what happened under the hood (I tried with or without breaking line in the Yaml, I tested different line separators for the Yaml file... This seems to be linked with the way base64 decoding is done.).